decodeDeploymentTx
